Auto truncation
Adding an exclamation point to a waypoint will make
it require a note to be present, and will skip to the
next waypoint if no note is detected inside the robot.

Even though the robot doesn't have enough time to make
it there normally, I would like to keep the 4th note in
the source side auto because of the possibility that the
other notes get stolen and truncated and it actually does
end up having enough time to make it there.